support generic task management
Demonstrate by providing `date` verb, which invokes the copy of `date`
on the host where the rcomp server is running.

This design does not provide means to inject data via stdin or
fs-visible files. It may be expanded toward supporting these cases,
but when generality becomes difficult, the preferred approach shall be
to introduce tool-specific interfaces.