support sending file for `ltl2ba` command
the implementation is generic, but code must be provided for commands
in order to decide which command-line arguments refer to files (and
thus, should have file data encoded).

N.B., while tool-agnostic inference could be used by comparing
command-line values with path names, this introduces the risk of
unintended information discloure, e.g., if a switch used by the tool
is coincidentally the same as a local file on disc.